chore: move merge into separate file
