+++ src/operators/index.ts
@@ -47,11 +47,12 @@ export { map } from '../internal/operators/map';
 export { mapTo } from '../internal/operators/mapTo';
 export { materialize } from '../internal/operators/materialize';
 export { max } from '../internal/operators/max';
-export { mergeWith, merge } from '../internal/operators/mergeWith';
+export { merge } from '../internal/operators/merge';
 export { mergeAll } from '../internal/operators/mergeAll';
 export { mergeMap, flatMap } from '../internal/operators/mergeMap';
 export { mergeMapTo } from '../internal/operators/mergeMapTo';
 export { mergeScan } from '../internal/operators/mergeScan';
+export { mergeWith } from '../internal/operators/mergeWith';
 export { min } from '../internal/operators/min';
 export { multicast } from '../internal/operators/multicast';
 export { observeOn } from '../internal/operators/observeOn';
